requirements.txt
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
chardet==3.0.4
constantly==15.1.0
defusedxml==0.5.0
Django==2.1.1
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
get==2018.11.19
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
mysqlclient==1.3.14
oauthlib==3.0.0
Pillow==5.2.0
post==2018.11.20
public==2018.11.20
pycrypto==2.6.1
pycryptodome==3.7.2
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
python-social-auth==0.3.6
python-weixin==0.4.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-core==3.0.0
Twisted==18.9.0
urllib3==1.24.1
userprofile==0.2.5
xmltodict==0.11.0
zope.interface==4.6.0
